requirements.txt
gradio==3.42.0
numpy==1.23.5
Pillow==10.0.0
tensorflow==2.10.0
